(fix)Update restaurants model
