+++ employees/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class EmployeeConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'employee'
+    name = 'employees'
