+++ restaurants/admin.py
@@ -1,0 +1,7 @@
+from django.contrib import admin
+
+from restaurants.models import Restaurant, Guest, Group
+
+# Register your models here.
+
+admin.site.register([Restaurant, Guest, Group])
