+++ restaurants/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class RestaurantConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'restaurant'
+    name = 'restaurants'
